poc_tagging/requirements.txt
github3.py==1.0.0a4
daiquiri==1.3.0
